+++ hooks/src/hooks/index.ts
@@ -1,4 +1,4 @@
 export { default as useCardNumber } from './useCardNumber/useCardNumber';
-export { default as useCvcNumber } from './useCVCNumber/useCVCNumber';
-export { default as useExpiryDate } from './useExpirationDate/useExpiryDate';
+export { default as useCVCNumber } from './useCVCNumber/useCVCNumber';
+export { default as useExpirationDate } from './useExpirationDate/useExpirationDate';
 export { default as usePassword } from './usePassword/usePassword';
